without narrowing conversations

+++ Integral/main.cpp
@@ -4,11 +4,11 @@
 #include <thread>
 
 static inline long double fun(long double x) {
-    return sin(1.0 / x);
+    return sinl(1.0L / x);
 }
 
 static inline long double max_first_dir(long double x) {
-    return fabs(2 / (x * x * x)) + fabs((long double) 1.0 / (x * x * x * x));
+    return fabsl(2 / (x * x * x)) + fabsl((long double) 1.0 / (x * x * x * x));
 }
 
 const long double finish_x = 1.0;
@@ -26,7 +26,7 @@ int main(int argc, char** argv) {
     int threads = (int) strtol(argv[1], nullptr, 10);
     const long double specific_error = strtod(argv[2], nullptr) / finish_x;
     printf("spec_error = %.100Lf\n", specific_error);
-    portion  = 1.0 / (threads * sqrt(specific_error));
+    portion  = static_cast<int>(1.0 / (threads * sqrtl(specific_error)));
     if (portion < 50) {
         portion = 50;
     }
@@ -35,7 +35,7 @@ int main(int argc, char** argv) {
         int statistic = 0;
         while (curr_x <= finish_x) {
             mutex_x.lock();
-            long double h = sqrt(
+            long double h = sqrtl(
                 specific_error * ((long double) 12.0) / max_first_dir(curr_x)
             );
             long double loc_x = curr_x;
@@ -47,7 +47,7 @@ int main(int argc, char** argv) {
             long double loc_sum = 0.0;
             const long double loc_start = loc_x;
             if (loc_x + h * portion >= finish_x) {
-                const int iters = (finish_x - loc_x) / h;
+                const int iters = static_cast<int>((finish_x - loc_x) / h);
                 for (int i = 0; i < iters; ++i) {
                     loc_sum += 0.5 * (fun(loc_x) + fun(loc_x + h));
                     loc_x += h;
